+++ Allegro/ParallelExpression.swift
@@ -16,7 +16,7 @@ public class ParallelExpression: CompoundExpression, ArrayLiteralConvertible {
     }
     
     public override var duration: Duration {
-        var result:Float = 0.0
+        var result:Double = 0.0
         for expression in self.expressions {
             result = max(result, expression.duration.length)
         }
